chatbot/chatbot-project-FastAPI/backend/requirements.txt
uvicorn
fastapi
httpx
sqlalchemy
pymysql
python-multipart
jose
python-jose
passlib
timestamp
dotenv
pyyaml